delete fixidx package
