+++ sql/analyzer/apply_indexes_from_outer_scope.go
@@ -18,8 +18,6 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/dolthub/go-mysql-server/sql/fixidx"
-
 	"github.com/dolthub/go-mysql-server/sql"
 	"github.com/dolthub/go-mysql-server/sql/expression"
 	"github.com/dolthub/go-mysql-server/sql/plan"
@@ -282,7 +280,7 @@ func getSubqueryIndexes(
 func tablesInScope(scope *plan.Scope) []string {
 	tables := make(map[string]bool)
 	for _, node := range scope.InnerToOuter() {
-		for _, col := range fixidx.Schemas(node.Children()) {
+		for _, col := range Schemas(node.Children()) {
 			tables[col.Source] = true
 		}
 	}
@@ -291,4 +289,13 @@ func tablesInScope(scope *plan.Scope) []string {
 		tableSlice = append(tableSlice, table)
 	}
 	return tableSlice
+}
+
+// Schemas returns the Schemas for the nodes given appended in to a single one
+func Schemas(nodes []sql.Node) sql.Schema {
+	var schema sql.Schema
+	for _, n := range nodes {
+		schema = append(schema, n.Schema()...)
+	}
+	return schema
 }
